scripts/requirements.txt
PyGithub>=2.1.1
openai>=1.55.0
requests>=2.31.0
